fix(base): preserve buffer ownership and value conversion semantics
Handle pooled allocation failure and aliased buffers, require shared
ownership for pool handles, and preserve stream data after resizing or
blocked writes. Correct shared-array traits and serialization dispatch.

Make Unicode conversion locale independent and hex output consistent
with field widths across helpers, parsers, and eproto. Correct Uint128
formatting and compiler-specific integer and branch-hint handling.

+++ cli/eproto/eproto_format.cc
@@ -84,13 +84,13 @@ class CustomFieldValuePrinter final : public google::protobuf::TextFormat::FastF
   void PrintInt32(int32_t val, google::protobuf::TextFormat::BaseTextGenerator* generator) const override {
     if (print_time_string && current_field_) {
       if (print_hex_string) {
-        generator->PrintString(vlink::Helpers::format_hex_number(static_cast<int64_t>(val)));
+        generator->PrintString(vlink::Helpers::format_hex_number(static_cast<uint64_t>(static_cast<uint32_t>(val))));
       } else {
         FastFieldValuePrinter::PrintInt32(val, generator);
       }
     } else {
       if (print_hex_string) {
-        generator->PrintString(vlink::Helpers::format_hex_number(static_cast<int64_t>(val)));
+        generator->PrintString(vlink::Helpers::format_hex_number(static_cast<uint64_t>(static_cast<uint32_t>(val))));
       } else {
         FastFieldValuePrinter::PrintInt32(val, generator);
       }
@@ -99,7 +99,7 @@ class CustomFieldValuePrinter final : public google::protobuf::TextFormat::FastF
 
   void PrintUInt32(uint32_t val, google::protobuf::TextFormat::BaseTextGenerator* generator) const override {
     if (print_hex_string) {
-      generator->PrintString(vlink::Helpers::format_hex_number(static_cast<int64_t>(val)));
+      generator->PrintString(vlink::Helpers::format_hex_number(static_cast<uint64_t>(val)));
     } else {
       FastFieldValuePrinter::PrintUInt32(val, generator);
     }
